Move more functionality into traits
