+++ src/main/scala/com/cognite/sdk/scala/v1/assets.scala
@@ -1,0 +1,26 @@
+package com.cognite.sdk.scala.v1
+
+import com.cognite.sdk.scala.common.WithId
+
+final case class Asset(
+    id: Long = 0,
+    path: Option[Seq[Long]] = None,
+    depth: Option[Long] = None,
+    name: String,
+    parentId: Option[Long] = None,
+    description: Option[String] = None,
+    metadata: Option[Map[String, String]] = None,
+    source: Option[String] = None,
+    externalId: Option[String] = None,
+    createdTime: Option[Long] = None,
+    lastUpdatedTime: Option[Long] = None
+) extends WithId[Long]
+
+final case class CreateAsset(
+    name: String,
+    parentId: Option[Long] = None,
+    description: Option[String] = None,
+    source: Option[String] = None,
+    externalId: Option[String] = None,
+    metadata: Option[Map[String, String]] = None
+)
