+++ src/main/scala/com/cognite/sdk/scala/v1/dataTypes.scala
@@ -1,3 +1,3 @@
-package com.cognite.sdk.scala.v1.resources
+package com.cognite.sdk.scala.v1
 
 final case class CogniteExternalId(id: String)
