+++ src/main/scala/com/cognite/sdk/scala/v1/events.scala
@@ -1,0 +1,30 @@
+package com.cognite.sdk.scala.v1
+
+import com.cognite.sdk.scala.common.WithId
+
+final case class Event(
+    id: Long = 0,
+    startTime: Option[Long] = None,
+    endTime: Option[Long] = None,
+    description: Option[String] = None,
+    `type`: Option[String] = None,
+    subtype: Option[String] = None,
+    metadata: Option[Map[String, String]] = None,
+    assetIds: Option[Seq[Long]] = None,
+    source: Option[String] = None,
+    externalId: Option[String] = None,
+    createdTime: Long = 0,
+    lastUpdatedTime: Long = 0
+) extends WithId[Long]
+
+final case class CreateEvent(
+    startTime: Option[Long] = None,
+    endTime: Option[Long] = None,
+    description: Option[String] = None,
+    `type`: Option[String] = None,
+    subtype: Option[String] = None,
+    metadata: Option[Map[String, String]] = None,
+    assetIds: Option[Seq[Long]] = None,
+    source: Option[String] = None,
+    externalId: Option[String] = None
+)
