+++ src/main/scala/com/cognite/sdk/scala/v1/resources/ReadWritableResourceV1.scala
@@ -1,33 +1,25 @@
 package com.cognite.sdk.scala.v1.resources
 
 import com.cognite.sdk.scala.common._
+import com.cognite.sdk.scala.v1.CogniteExternalId
 import com.softwaremill.sttp._
 import com.softwaremill.sttp.circe._
-import io.circe.generic.auto._
 import io.circe.{Decoder, Encoder}
 
 abstract class ReadWritableResourceV1[R: Decoder, W: Decoder: Encoder, F[_]](
     implicit auth: Auth
 ) extends ReadWritableResourceWithRetrieve[R, W, F, Id, CogniteId, Long]
     with DeleteByIdsV1[R, W, F, Id, CogniteId, Long]
-    with DeleteByExternalIdsV1[F, String] {}
+    with DeleteByExternalIdsV1[F] {}
 
-trait DeleteByExternalIds[F[_], ExternalId] {
-  def deleteByExternalIds(externalIds: Seq[String])(
-      implicit sttpBackend: SttpBackend[F, _],
-      errorDecoder: Decoder[CdpApiError[CogniteId]],
-      itemsEncoder: Encoder[Items[ExternalId]]
-  ): F[Response[Unit]]
-}
-
-trait DeleteByExternalIdsV1[F[_], ExternalId]
+trait DeleteByExternalIdsV1[F[_]]
     extends RequestSession
     with BaseUri
-    with DeleteByExternalIds[F, ExternalId] {
+    with DeleteByExternalIds[F, CogniteId, CogniteExternalId] {
   def deleteByExternalIds(externalIds: Seq[String])(
       implicit sttpBackend: SttpBackend[F, _],
       errorDecoder: Decoder[CdpApiError[CogniteId]],
-      itemsEncoder: Encoder[Items[ExternalId]]
+      itemsEncoder: Encoder[Items[CogniteExternalId]]
   ): F[Response[Unit]] = {
     implicit val errorOrUnitDecoder: Decoder[Either[CdpApiError[CogniteId], Unit]] =
       EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Unit]
