+++ src/main/scala/com/cognite/sdk/scala/v1/resources/events.scala
@@ -1,35 +1,9 @@
 package com.cognite.sdk.scala.v1.resources
 
-import com.cognite.sdk.scala.common.{Auth, WithId}
+import com.cognite.sdk.scala.common.Auth
+import com.cognite.sdk.scala.v1.{CreateEvent, Event}
 import com.softwaremill.sttp._
 import io.circe.generic.auto._
-
-final case class Event(
-    id: Long = 0,
-    startTime: Option[Long] = None,
-    endTime: Option[Long] = None,
-    description: Option[String] = None,
-    `type`: Option[String] = None,
-    subtype: Option[String] = None,
-    metadata: Option[Map[String, String]] = None,
-    assetIds: Option[Seq[Long]] = None,
-    source: Option[String] = None,
-    externalId: Option[String] = None,
-    createdTime: Long = 0,
-    lastUpdatedTime: Long = 0
-) extends WithId[Long]
-
-final case class CreateEvent(
-    startTime: Option[Long] = None,
-    endTime: Option[Long] = None,
-    description: Option[String] = None,
-    `type`: Option[String] = None,
-    subtype: Option[String] = None,
-    metadata: Option[Map[String, String]] = None,
-    assetIds: Option[Seq[Long]] = None,
-    source: Option[String] = None,
-    externalId: Option[String] = None
-)
 
 class Events[F[_]](project: String)(implicit auth: Auth)
     extends ReadWritableResourceV1[Event, CreateEvent, F]
