+++ src/main/scala/com/cognite/sdk/scala/v1/resources/threeD.scala
@@ -1,22 +1,16 @@
 package com.cognite.sdk.scala.v1.resources
 
 import com.cognite.sdk.scala.common._
+import com.cognite.sdk.scala.v1.{
+  CreateThreeDModel,
+  CreateThreeDRevision,
+  ThreeDModel,
+  ThreeDRevision
+}
 import com.softwaremill.sttp._
 import com.softwaremill.sttp.circe._
 import io.circe.{Decoder, Encoder}
 import io.circe.generic.auto._
-
-final case class ThreeDModel(
-    name: String,
-    id: Long = 0,
-    createdTime: Option[Long] = None,
-    metadata: Option[Map[String, String]] = None
-) extends WithId[Long]
-
-final case class CreateThreeDModel(
-    name: String,
-    metadata: Option[Map[String, String]] = None
-)
 
 class ThreeDModels[F[_]](project: String)(implicit auth: Auth)
     extends ReadWritableResource[ThreeDModel, CreateThreeDModel, F, Id, CogniteId, Long]
@@ -29,7 +23,9 @@ class ThreeDModels[F[_]](project: String)(implicit auth: Auth)
       implicit sttpBackend: SttpBackend[F, _],
       errorDecoder: Decoder[CdpApiError[CogniteId]],
       itemsEncoder: Encoder[Items[CogniteId]]
-  ): F[Response[Unit]] =
+  ): F[Response[Unit]] = {
+    implicit val errorOrUnitDecoder: Decoder[Either[CdpApiError[CogniteId], Unit]] =
+      EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Unit]
     // TODO: group deletes by max deletion request size
     //       or assert that length of `ids` is less than max deletion request size
     request
@@ -42,34 +38,9 @@ class ThreeDModels[F[_]](project: String)(implicit auth: Auth)
         case Right(Right(_)) => ()
       }
       .send()
+  }
+
 }
-
-final case class Camera(
-    target: Option[Array[Double]],
-    position: Option[Array[Double]]
-)
-
-final case class ThreeDRevision(
-    id: Long,
-    fileId: Long,
-    published: Boolean,
-    rotation: Option[Array[Double]] = None,
-    camera: Option[Camera] = None,
-    status: String,
-    metadata: Option[Map[String, String]] = None,
-    thumbnailThreedFileId: Option[Long] = None,
-    thumbnailURL: Option[String] = None,
-    assetMappingCount: Long,
-    createdTime: Long
-) extends WithId[Long]
-
-final case class CreateThreeDRevision(
-    published: Boolean,
-    rotation: Option[Array[Double]] = None,
-    metadata: Option[Map[String, String]] = None,
-    camera: Option[Camera] = None,
-    fileId: Long
-)
 
 class ThreeDRevisions[F[_]](project: String, modelId: Long)(
     implicit auth: Auth
@@ -85,13 +56,13 @@ class ThreeDRevisions[F[_]](project: String, modelId: Long)(
   override val baseUri =
     uri"https://api.cognitedata.com/api/v1/projects/$project/3d/models/$modelId/revisions"
 
-  implicit val errorOrUnitDecoder: Decoder[Either[CdpApiError[CogniteId], Unit]] =
-    EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Unit]
   def deleteByIds(ids: Seq[Long])(
       implicit sttpBackend: SttpBackend[F, _],
       errorDecoder: Decoder[CdpApiError[CogniteId]],
       itemsEncoder: Encoder[Items[CogniteId]]
-  ): F[Response[Unit]] =
+  ): F[Response[Unit]] = {
+    implicit val errorOrUnitDecoder: Decoder[Either[CdpApiError[CogniteId], Unit]] =
+      EitherDecoder.eitherDecoder[CdpApiError[CogniteId], Unit]
     // TODO: group deletes by max deletion request size
     //       or assert that length of `ids` is less than max deletion request size
     request
@@ -104,6 +75,8 @@ class ThreeDRevisions[F[_]](project: String, modelId: Long)(
         case Right(Right(_)) => ()
       }
       .send()
+  }
+
 }
 
 final case class ThreeDAssetMapping(
