+++ src/main/scala/com/cognite/sdk/scala/v1/resources/timeSeries.scala
@@ -1,34 +1,9 @@
 package com.cognite.sdk.scala.v1.resources
 
-import com.cognite.sdk.scala.common.{Auth, WithId}
+import com.cognite.sdk.scala.common.Auth
+import com.cognite.sdk.scala.v1.{CreateTimeSeries, TimeSeries}
 import com.softwaremill.sttp._
 import io.circe.generic.auto._
-
-final case class TimeSeries(
-    name: String,
-    isString: Boolean = false,
-    metadata: Option[Map[String, String]] = None,
-    unit: Option[String] = None,
-    assetId: Option[Long] = None,
-    isStep: Boolean = false,
-    description: Option[String] = None,
-    securityCategories: Option[Seq[Long]] = None,
-    id: Long = 0,
-    externalId: Option[String] = None,
-    createdTime: Long = 0,
-    lastUpdatedTime: Long = 0
-) extends WithId[Long]
-
-final case class CreateTimeSeries(
-    name: String,
-    isString: Boolean = false,
-    metadata: Option[Map[String, String]] = None,
-    unit: Option[String] = None,
-    assetId: Option[Long] = None,
-    isStep: Boolean = false,
-    description: Option[String] = None,
-    securityCategories: Option[Seq[Long]] = None
-)
 
 class TimeSeriesResource[F[_]](project: String)(
     implicit auth: Auth
