+++ src/main/scala/com/cognite/sdk/scala/v1/timeSeries.scala
@@ -1,0 +1,29 @@
+package com.cognite.sdk.scala.v1
+
+import com.cognite.sdk.scala.common.WithId
+
+final case class TimeSeries(
+    name: String,
+    isString: Boolean = false,
+    metadata: Option[Map[String, String]] = None,
+    unit: Option[String] = None,
+    assetId: Option[Long] = None,
+    isStep: Boolean = false,
+    description: Option[String] = None,
+    securityCategories: Option[Seq[Long]] = None,
+    id: Long = 0,
+    externalId: Option[String] = None,
+    createdTime: Long = 0,
+    lastUpdatedTime: Long = 0
+) extends WithId[Long]
+
+final case class CreateTimeSeries(
+    name: String,
+    isString: Boolean = false,
+    metadata: Option[Map[String, String]] = None,
+    unit: Option[String] = None,
+    assetId: Option[Long] = None,
+    isStep: Boolean = false,
+    description: Option[String] = None,
+    securityCategories: Option[Seq[Long]] = None
+)
