+++ src/test/scala/com/cognite/sdk/scala/v1/AssetsTest.scala
@@ -1,15 +1,14 @@
 package com.cognite.sdk.scala.v1
 
-import com.cognite.sdk.scala.common.{ReadableResourceBehaviors, SdkTest, WritableResourceBehaviors}
-import com.cognite.sdk.scala.v1.resources.{Asset, CreateAsset}
+import com.cognite.sdk.scala.common.{ReadBehaviours, SdkTest, WritableBehaviors}
 
-class AssetsTest extends SdkTest with ReadableResourceBehaviors with WritableResourceBehaviors {
+class AssetsTest extends SdkTest with ReadBehaviours with WritableBehaviors {
   private val client = new GenericClient()(auth, sttpBackend)
   private val idsThatDoNotExist = Seq(999991L, 999992L)
 
-  it should behave like readableResource(client.assets)
-  it should behave like readableResourceWithRetrieve(client.assets, idsThatDoNotExist, supportsMissingAndThrown = true)
-  it should behave like writableResource(
+  it should behave like readable(client.assets)
+  it should behave like readableWithRetrieve(client.assets, idsThatDoNotExist, supportsMissingAndThrown = true)
+  it should behave like writable(
     client.assets,
     Seq(Asset(name = "scala-sdk-read-example-1"), Asset(name = "scala-sdk-read-example-2")),
     Seq(CreateAsset(name = "scala-sdk-create-example-1"), CreateAsset(name = "scala-sdk-create-example-2")),
